refactor: annotate the last untyped definitions, and test the untested modules
mypy now runs with disallow_untyped_defs, warn_return_any, warn_unreachable
and no_implicit_optional. Switching those on found three real things rather
than just missing annotations: db_to_linear and _request both widened to Any
through a concrete return type, and cmd_pack was being handed None for a
parameter typed as a PipelineConfig. That last one is honest now: pack,
presets and verify-upload run before a config exists and say so in the
signature, instead of the call site lying.

doctor, verify and voicelab had no tests at all, which is 650 lines of
user-facing code including two modules that do network I/O. 33 cases now,
covering the things that would be silent if wrong:

- verify: tar members that try to escape the extraction directory, since the
  archive is downloaded from the internet; every field that should falsify the
  verdict actually doing so; a byte-size mismatch against the local pack, which
  is the check that catches a placeholder surviving an upload.
- voicelab: provenance, where an unrecognised or missing category has to read
  as UNVERIFIED rather than defaulting to something reassuring.
- doctor: the three key states, because a placeholder reported as present is
  worse than one reported as missing.

Docs: the README told people to set OPENAI_API_KEY to "sk-...", which is
exactly the placeholder that later got set for real and cost a debugging
session, so it no longer looks pasteable. Adds the --accept-voice-terms flag
the first build needs, the pip install route to the `wvs` command, and an index
of the eight docs. CONTRIBUTING now lists four CI checks rather than three.

+++ waze_voice/media.py
@@ -246,7 +246,8 @@ RESET_TIMELINE = "asetpts=N/SR/TB"
 
 
 def db_to_linear(db: float) -> float:
-    return 10.0 ** (db / 20.0)
+    # float(...) because ** widens to Any under a strict checker.
+    return float(10.0 ** (db / 20.0))
 
 
 def trim_silence_chain(threshold_db: float) -> list[str]:
